proto: extract ReviewSurface, add /scan route (issue #259)
Extract receipt review UI (ReviewSurface) as presentation-only component,
enabling reuse across sheet mounts and future flows (chat #254). Add /scan
route with full upload state machine, giving mobile users more room to
review items than a bottom sheet allows.

Key changes:
- ReviewSurface: new component, purely presentation (no fetch/routing)
- /scan/page.tsx: route owning upload→processing→review→confirm flow
- ScanTab: updated to use ReviewSurface (still sheet container)
- Tests updated to target ReviewSurface
